SFEperl-devel-ppport: bump to 3.36 and build packages for perl-522 and perl-526
--HG--
branch : spec_SFEperl-devel-ppport

+++ specs/SFEperl-devel-ppport.spec
@@ -4,8 +4,12 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build526 %( if [ -x /usr/perl5/5.26/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
+
 %define include_executable 0
+%define install_to_site_dir 0
 
 %define cpan_name Devel-PPPort
 %define sfe_cpan_name devel-ppport
@@ -14,8 +18,8 @@
 Summary:               Perl/Pollution/Portability
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               3.32
-IPS_component_version: 3.32
+Version:               3.36
+IPS_component_version: 3.36
 License:               perl_5
 URL:                   https://metacpan.org/pod/Devel::PPPort
 Source0:               http://cpan.metacpan.org/authors/id/W/WO/WOLFSAGE/Devel-PPPort-%{version}.tar.gz
@@ -30,6 +34,8 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Perl/Pollution/Portability
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+%if %{enable_test}
+%endif
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 
@@ -56,6 +62,8 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Perl/Pollution/Portability
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+%if %{enable_test}
+%endif
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 
@@ -69,6 +77,9 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Perl/Pollution/Portability
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+%endif
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 
@@ -76,52 +87,87 @@ Requires:         library/perl-5/%{ips_cpan_name}
 Perl/Pollution/Portability
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Perl/Pollution/Portability
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+%if %{enable_test}
+%endif
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 
-%description 520
+%description 522
 Perl/Pollution/Portability
 %endif
 
+%if %{build526}
+%package 526
+IPS_package_name: library/perl-5/%{ips_cpan_name}-526
+Summary:          Perl/Pollution/Portability
+BuildRequires:    runtime/perl-526 = *
+BuildRequires:    library/perl-5/extutils-makemaker-526
+%if %{enable_test}
+%endif
+Requires:         runtime/perl-526 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+
+%description 526
+Perl/Pollution/Portability
+%endif
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
     perl_ver=$1
     test=$2
-    bindir="/usr/perl5/${perl_ver}/bin"
+    perl_dir_prefix="/usr/perl5/${perl_ver}"
+    bindir="${perl_dir_prefix}/bin"
     vendor_dir="/usr/perl5/vendor_perl/${perl_ver}"
+    site_dir="/usr/perl5/site_perl/${perl_ver}"
 
     export PERL5LIB=${vendor_dir}
+%if %{install_to_site_dir}
+    perl_libdir="${site_dir}"
+%else
+    perl_libdir="${vendor_dir}"
+%endif
+
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
-                   LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+                   LIB=${perl_libdir}
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
 build_with_build.pl_for() {
     perl_ver=$1
     test=$2
-    bindir="/usr/perl5/${perl_ver}/bin"
+    perl_dir_prefix="/usr/perl5/${perl_ver}"
+    bindir="${perl_dir_prefix}/bin"
     vendor_dir="/usr/perl5/vendor_perl/${perl_ver}"
+    site_dir="/usr/perl5/site_perl/${perl_ver}"
 
+%if %{install_to_site_dir}
+    installdir='site'
+%else
+    installdir='vendor'
+%endif
     export PERL5LIB=${vendor_dir}
     ${bindir}/perl Build.PL \
-                   --installdirs vendor \
+                   --installdirs ${installdir} \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -138,7 +184,7 @@ modify_bin_dir() {
     then
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -156,7 +202,11 @@ modify_man_dir() {
             mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
             rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
         fi
-        rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
     fi
 }
 
@@ -191,8 +241,12 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
+%endif
+
+%if %{build526}
+build_for 5.26
 %endif
 
 %install
@@ -217,7 +271,11 @@ rm -rf %{buildroot}
 %files 584
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.8.4
+%else
 /usr/perl5/vendor_perl/5.8.4
+%endif
 %if %{include_executable}
 /usr/perl5/5.8.4
 %endif
@@ -227,7 +285,11 @@ rm -rf %{buildroot}
 %files 510
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.10
+%else
 /usr/perl5/vendor_perl/5.10
+%endif
 %if %{include_executable}
 /usr/perl5/5.1.0
 %endif
@@ -237,7 +299,11 @@ rm -rf %{buildroot}
 %files 512
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.12
+%else
 /usr/perl5/vendor_perl/5.12
+%endif
 %if %{include_executable}
 /usr/perl5/5.12
 %endif
@@ -247,24 +313,47 @@ rm -rf %{buildroot}
 %files 516
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.16
+%else
 /usr/perl5/vendor_perl/5.16
+%endif
 %if %{include_executable}
 /usr/perl5/5.16
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.22
+%else
+/usr/perl5/vendor_perl/5.22
+%endif
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
 
+%if %{build526}
+%files 526
+%defattr(0755,root,bin,-)
+%dir %attr (0755, root, sys) /usr
+%if %{install_to_site_dir}
+/usr/perl5/site_perl/5.26
+%else
+/usr/perl5/vendor_perl/5.26
+%endif
+%if %{include_executable}
+/usr/perl5/5.26
+%endif
+%endif
 
 %changelog
+* Fri Mar 16 2018 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 3.36 and build packages for perl-522 and perl-526
 * Sat Nov 14 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 3.32 and build packages for perl-510, perl-516 and perl-520
 * Tue Nov 05 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
